requirements.txt
agno
yfinance
openai 
duckduckgo-search 
newspaper4k 
lxml_html_clean
groq
python-dotenv
streamlit
